Sonarcloud issue log (#14)
* Structured binding should be used
C++17 introduced structured binding, a syntax that initializes multiple entities by elements or members of an object.

* Implicit casts should not lower precision
implicit conversion loses integer precision: 'const uint64_t' (aka 'const unsigned long') to 'uint32_t' (aka 'unsigned int')

* Use '=default' instead of the default implementation of this special member functions.
Special member function should not be defined unless a non standard behavior is required

* Member functions that don't mutate their objects should be declared 'const'

* Pointer and reference parameters should be 'const' if the corresponding object is not modified

* Pointer and reference local variables should be 'const' if the corresponding object is not modified

+++ src/drm.h
@@ -109,7 +109,7 @@ class PropertyValue {
   PropertyValue(uint32_t id, uint64_t value) : id_(id), value_(value) {}
 
   uint32_t id() const { return id_; }
-  uint32_t value() const { return value_; }
+  uint64_t value() const { return value_; }
 
  private:
   uint32_t id_;
